pyproject.toml
[project]
name = "kiosk-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["aerich>=0.8.2", "pydantic>=2.11.4", "python-dotenv>=1.1.0", "sanic[ext]>=25.3.0", "tomlkit>=0.13.2", "tortoise-orm[asyncpg]>=0.25.0"]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.aerich]
tortoise_orm = "app.config.TORTOISE_ORM_CONFIG"
location = "./migrations"
src_folder = "./."

[dependency-groups]
dev = ["ruff>=0.11.8"]
